--- a/clovers/src/draw.rs
+++ b/clovers/src/draw.rs
@@ -14,9 +14,6 @@ pub fn draw(
     gamma: Float,
     scene: Scene,
 ) -> Vec<Color> {
-    let span = span!(Level::TRACE, "draw");
-    let _enter = span.enter();
-
     // Progress bar
     let pixels = (width * height) as u64;
     let bar = ProgressBar::new(pixels);
@@ -32,6 +29,7 @@ pub fn draw(
         .par_iter_mut()
         .enumerate()
         .for_each(|(index, pixel)| {
+            let span = span!(Level::TRACE, "draw", index, ?pixel);
             let _enter = span.enter();
             let x = index % width as usize;
             let y = index / width as usize;